multiplies two positive numbers

+++ more_malloc_free/101-mul.c
@@ -39,7 +39,7 @@ int _strlen(char *s)
 char *big_multiply(char *s1, char *s2)
 {
 	char *r;
-	int l1, l2, a, b, c, x;
+	long int l1, l2, a, b, c, x;
 
 	l1 = _strlen(s1);
 	l2 = _strlen(s2);
@@ -90,7 +90,7 @@ char *big_multiply(char *s1, char *s2)
 int main(int argc, char **argv)
 {
 	char *r;
-	int a, c, x;
+	long int a, c, x;
 
 	if (argc != 3)
 		printf("Error\n"), exit(98);
